+++ CHANGES.rst
@@ -14,6 +14,8 @@ New Features:
   restart* and *any restart* values to restart page numbering
 * The new *continue* page number format makes it more explicit when to not
   restart page numbering.
+* Setting the *base* for a style to ``NEXT_STYLE`` proceeds to look up style
+  attributes in the next matching style if they are undefined.
 * If the *RINOH_NO_CACHE* environment variable is set, the references cache
   (.rtc file) won't be loaded nor saved. This is mostly useful for testing.